requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.4
async-timeout==3.0.1
attrs==21.2.0
backports.entry-points-selectable==1.1.0
black==21.7b0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.1
click==8.0.1
colorama==0.4.4
discord.py==1.7.3
distlib==0.3.2
filelock==3.0.12
idna==3.2
multidict==5.1.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.0.2
pygame==2.0.1
python-dotenv==0.18.0
regex==2021.7.6
requests==2.26.0
Send2Trash==1.7.1
six==1.16.0
tomli==1.0.4
typing-extensions==3.10.0.0
urllib3==1.26.6
virtualenv==20.5.0
yarl==1.6.3